misc: use description asset on course overview pg
Check if T_OG_DESCRIPTION asset exists for a module and use that
when rendering course modules on course overview page.

Otherwise we leave it blank.

This is probably not the best solution. Should guarantee that
T_OG_DESCRIPTION should exist for these pages (so we can improve
accessibility, shareability, and SEO), or come up with some
other reasonable placeholder if a description doesn't exist