feat: add nav from submodule to next/prev submods
Navigation between submodules is here!

Implemented this with a custom GraphQL query and some filtering logic
to get the next/prev nodes in sequence based on connections.

To customize the order of submodules in the future, we can add
edge weights and make sure to sort by edgeweights before we do this
pseudo-pagination logic